fixed compiler error in speciesgraph
git-svn-id: file:///home/gumz3/Desktop/svnAugustusRepo/svnArchiveFile/svn/augustus/trunks@761 946bced8-b911-4d30-a262-e0a769a8c5ad

Former-commit-id: 07377e32a6191c7c9e0a2595e5f9bf9da1f4e542

+++ include/speciesgraph.hh
@@ -115,7 +115,7 @@ private:
     Node* getAuxilaryNode(NodeType type, int pos, map<string,Node*> &auxiliaryNodes) const; // get an auxilary node (returns NULL if node does not exists)
     void addAuxilaryEdge(Node *pred, Node *succ); // add an auxilary edge to the graph
     Node* addAuxNodeToLine(NodeType type, int pos, vector< vector<Node*> >&neutralLines); // add an auxilary node to one of the neutral lines if it does not exist already
-    NodeType getPredType(StateType type, double begin, double end) ;  // get the type of gene feature that precedes an exon
+    NodeType getPredType(StateType type, int begin, int end) ;  // get the type of gene feature that precedes an exon
     NodeType getSuccType(StateType type) ;    // get the type of gene feature that succeeds an exon 
     list<NodeType> getPredTypes(Node *node) ;  // same as above, but includes stop codon types 
     list<NodeType> getSuccTypes(Node *node) ;    // same as above, but includes stop codon types
